requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
alembic==1.12.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
blinker==1.8.2
click==8.1.7
colorama==0.4.6
flake8==5.0.4
Flask==3.0.2
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
PyYAML==6.0.1
frozenlist==1.4.1
greenlet==3.0.3
idna==3.8
importlib_metadata==7.1.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.3
MarkupSafe==2.1.5
mccabe==0.7.0
multidict==6.0.5
packaging==24.0
pluggy==1.5.0
pillow==10.4.0
py==1.11.0
pycodestyle==2.9.1
pyflakes==2.5.0
pytest==7.1.3
pytest_aiohttp==1.0.5
pytest-asyncio==0.23.4
pytest-env==0.6.2
pytest-tornasync==0.6.0.post2
python-dateutil==2.8.2
python-dotenv==1.0.0
six==1.16.0
SQLAlchemy==2.0.21
tomli==2.0.1
typing_extensions==4.11.0
Werkzeug==3.0.0
WTForms==3.0.1
yarl==1.9.9
zipp==3.18.1
